Passage à git

+++ src/net/enteranamestudio/project/network/packets/PacketInitPlayer.java
@@ -1,10 +1,9 @@
 package net.enteranamestudio.project.network.packets;
 
 public class PacketInitPlayer {
-	public int id;
+	public byte id;
 	public String name;
 	public int player;
-	
-	public int x;
-	public int y;
+	public short x;
+	public short y;
 }
